feat: Pasar funciones a utilidades
